fixed compiler warnings

+++ BasicOpenGLEngine/BasicOpenGLEngine/Light.cpp
@@ -12,7 +12,7 @@ Light::Light(GLfloat shadow_width, GLfloat shadow_height,
 					  GLfloat red, GLfloat green, GLfloat blue,	
 					  GLfloat a_intensity, GLfloat d_intensity) {
 	shadow_map = new ShadowMap();
-	shadow_map->init(shadow_width, shadow_height);
+	shadow_map->init((GLuint)shadow_width, (GLuint)shadow_height);
 
 	color = glm::vec3(red, green, blue);
 	ambient_intensity = a_intensity;
